Rename firmware_updater to _firmware_updater
